+++ packages/zowe-explorer/CHANGELOG.md
@@ -6,14 +6,20 @@ All notable changes to the "vscode-extension-for-zowe" extension will be documen
 
 ### New features and enhancements
 
+### Bug fixes
+
+## `2.10.0`
+
+### New features and enhancements
+
 - Added call to callback if defined by extenders when a change to the team config profile is made. [#2385](https://github.com/zowe/vscode-extension-for-zowe/issues/2385)
-- Replaced `keytar` dependency with `keyring` module from [`@zowe/secrets-for-zowe-sdk`](https://github.com/zowe/zowe-cli/tree/master/packages/secrets). [#2358](https://github.com/zowe/vscode-extension-for-zowe/issues/2358)
+- Replaced `keytar` dependency with `keyring` module from [`@zowe/secrets-for-zowe-sdk`](https://github.com/zowe/zowe-cli/tree/master/packages/secrets). [#2358](https://github.com/zowe/vscode-extension-for-zowe/issues/2358) [#2348](https://github.com/zowe/vscode-extension-for-zowe/issues/2348)
 - Added "Edit Attributes" option for USS files and folders. [#2254](https://github.com/zowe/vscode-extension-for-zowe/issues/2254)
 
 ### Bug fixes
 
 - Fix the USS refresh icon (replacing "download" with "refresh")
-- Fix error for Theia check when token authentication returns 401.
+- Fix error for Theia check when token authentication returns 401. [#2407](https://github.com/zowe/vscode-extension-for-zowe/issues/2407)
 
 ## `2.9.2`